+++ src/Zope2/Startup/tests/test_schema.py
@@ -208,6 +208,19 @@ class StartupTestCase(unittest.TestCase):
         self.assertEqual(conf.databases[0].config.connection_class.__name__,
                          'LowConflictConnection')
 
+    def test_max_conflict_retries_default(self):
+        conf, handler = self.load_config_text("""\
+            instancehome <<INSTANCE_HOME>>
+            """)
+        self.assertEqual(conf.max_conflict_retries, 3)
+
+    def test_max_conflict_retries_explicit(self):
+        conf, handler = self.load_config_text("""\
+            instancehome <<INSTANCE_HOME>>
+            max-conflict-retries 15
+            """)
+        self.assertEqual(conf.max_conflict_retries, 15)
+
 def test_suite():
     return unittest.makeSuite(StartupTestCase)